+++ src/fmri/glm.c
@@ -87,8 +87,8 @@
  */
 
 #define MALLOC_FAILURE(num,type)\
-  { fprintf(stderr,"%s:%d: %s: unable to allocate %d %ss\n",\
-            __FILE__,__LINE__,__func__,num,#type); exit(-1); }
+  { fprintf(stderr,"%s:%d: %s: unable to allocate %ld %ss\n",\
+            __FILE__,__LINE__,__func__,(long)num,#type); exit(-1); }
 
 #define CHECK_DEBUG(r) (r->get(r,GLM_DEBUG))
